tweaking metrics

+++ src/benchmark/run_all_benchmarks.py
@@ -3,6 +3,7 @@ import time
 import subprocess
 import csv
 from datetime import datetime
+from os import wait
 from pathlib import Path
 import uuid
 import yaml
@@ -39,7 +40,7 @@ cvc5_configs = [
 ]
 
 
-def run_command_with_metrics(command, sampling_interval=0.1, timeout=None):
+def run_command_with_metrics(command, sampling_interval=1, timeout=None):
     start_time = time.time()
     print(f"Running command: {command}")
     process = psutil.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
@@ -50,16 +51,14 @@ def run_command_with_metrics(command, sampling_interval=0.1, timeout=None):
     try:
         while process.poll() is None:
             try:
-                cpu_percent = process.cpu_percent()
-                memory_info = process.memory_info()
+                all_cpu = psutil.cpu_percent(interval=sampling_interval, percpu=True)
+                cpu_usage.append(all_cpu)
 
-                cpu_usage.append(cpu_percent)
-                memory_usage.append(memory_info.rss / 1024 / 1024)
+                memory_info = process.memory_info()
+                memory_usage.append(memory_info.rss / (1024 * 1024))  # megabytes
 
             except psutil.NoSuchProcess:
                 break
-
-            time.sleep(sampling_interval)
 
             if timeout and (time.time() - start_time) > timeout:
                 process.terminate()
@@ -70,15 +69,17 @@ def run_command_with_metrics(command, sampling_interval=0.1, timeout=None):
         end_time = time.time()
         elapsed_time = end_time - start_time
 
-        avg_cpu = sum(cpu_usage) / len(cpu_usage) if cpu_usage else 0
+        avg_all_cpu = [sum(core) / len(core) for core in zip(*cpu_usage)] if cpu_usage else []
+        avg_cpu_percentage = sum(avg_all_cpu) / len(avg_all_cpu) if avg_all_cpu else 0
+
         max_memory = max(memory_usage) if memory_usage else 0
 
         print(f"Command finished: {command}")
-        return process.returncode, elapsed_time, stdout, stderr, avg_cpu, max_memory
+        return process.returncode, elapsed_time, stdout, stderr, avg_cpu_percentage, max_memory, avg_all_cpu
 
     except Exception as e:
         print(f"An error occurred while running the command: {e}")
-        return None, None, b"", str(e).encode(), None, None
+        return None, None, b"", str(e).encode(), None, None, None
 
 
 def run_command(command):
@@ -107,7 +108,7 @@ def run_cvc5_experiments(timeout_ms=30000):
     run_id = str(uuid.uuid4())
     run_datetime = datetime.now().isoformat()
 
-    fieldnames = ["run_id", "run_datetime", "solver", "config", "file", "return_code", "time", "avg_cpu_usage", "max_memory_usage", "stdout", "stderr"]
+    fieldnames = ["run_id", "run_datetime", "solver", "config", "file", "return_code", "time", "avg_cpu_percentage", "max_memory_usage","avg_cpu_usage", "stdout", "stderr"]
 
     sygus_files = list(Path.joinpath(project_root, sygus_dir).glob("*.sl"))
     for sygus_file in sygus_files:
@@ -117,7 +118,8 @@ def run_cvc5_experiments(timeout_ms=30000):
             results = []
             print(f"Running cvc5 with config: {config}")
             command = f"cvc5 --tlimit={timeout_ms} {config} {sygus_file}"
-            retcode, elapsed_time, stdout, stderr, avg_cpu, max_memory = run_command_with_metrics(command, timeout=timeout_ms/1000)
+            (retcode, elapsed_time, stdout, stderr, avg_cpu_percentage, max_memory,
+             avg_all_cpu) = run_command_with_metrics(command, timeout=timeout_ms/1000)
             results.append({
                 "run_id": run_id,
                 "run_datetime": run_datetime,
@@ -126,8 +128,9 @@ def run_cvc5_experiments(timeout_ms=30000):
                 "file": str(sygus_file),
                 "return_code": retcode,
                 "time": elapsed_time,
-                "avg_cpu_usage": avg_cpu,
+                "avg_cpu_percentage": avg_cpu_percentage,
                 "max_memory_usage": max_memory,
+                "avg_cpu_usage": json.dumps(avg_all_cpu),
                 "stdout": stdout.decode("utf-8"),
                 "stderr": stderr.decode("utf-8"),
             })
@@ -136,19 +139,34 @@ def run_cvc5_experiments(timeout_ms=30000):
     print(f"All cvc5 finished")
 
 
-def run_pysynthlab_experiments():
+def run_pysynthlab_experiments(timeout_seconds: int = 30):
     run_id = str(uuid.uuid4())
     run_datetime = datetime.now().isoformat()
 
-    fieldnames = ["run_id", "run_datetime", "solver", "config", "file", "return_code", "time", "stdout", "stderr"]
-    fieldnames += ["candidates_generated", "candidates_pruned", "iterations", "solution_height", "solution_complexity", "best_partial_score", "grammar_size", "solution_found", "avg_cpu_usage", "max_cpu_usage", "avg_per_core_cpu_usage", "max_per_core_cpu_usage", "max_memory_usage", "solution_space_coverage", "solver_calls", "avg_solver_time", "unique_patterns", "pattern_reuse_ratio"]
+    fieldnames = [
+        "run_id", "run_datetime", "solver", "config", "file", "return_code", "time", "avg_cpu_percentage", 
+        "max_memory_usage", "avg_all_cpu", "stdout", "stderr",
+        "candidates_generated",
+        "candidates_pruned",
+        "iterations",
+        "time_spent",
+        "solution_height",
+        "solution_complexity",
+        "best_partial_score",
+        "grammar_size",
+        "solution_found",
+        "solution_space_coverage",
+        "solver_calls",
+        "unique_patterns",
+        "pattern_reuse_ratio",
+    ]
 
     sygus_files = list(Path.joinpath(project_root, sygus_dir).glob("*.sl"))
     for sygus_file in sygus_files:
-        results = []
         print(f"Processing SyGuS file: {sygus_file}")
 
         for config_file in py_synth_lab_solver_configs:
+            results = []
             print(f"Running PySynthLab with config file: {config_file}")
             config = ConfigManager.load_yaml(config_file)
             if config is None:
@@ -164,6 +182,7 @@ def run_pysynthlab_experiments():
                 problem=sygus_file.stem,
                 root=project_root
             )
+            config['synthesis_parameters']['timeout'] = timeout_seconds
             benchmark_name = Path(config_file).stem
             temp_config_file = f"temp_{benchmark_name}_config.yaml"
             with open(temp_config_file, 'w') as f:
@@ -171,10 +190,20 @@ def run_pysynthlab_experiments():
 
             metrics_file = config['logging']['metrics_file']
             command = str(project_root) + f"/pysynthlab/bin/python -m src.runner --config {Path.joinpath(Path.cwd(), temp_config_file)} {sygus_file}"
-            retcode, elapsed_time, stdout, stderr = run_command(command)
+            retcode, elapsed_time, stdout, stderr, avg_cpu_percentage, max_memory, avg_all_cpu = run_command_with_metrics(command, timeout=timeout_seconds)
 
-            with open(metrics_file, 'r') as f:
-                metrics = json.load(f)
+            max_wait = 5
+            wait_time = 0
+            while not Path(metrics_file).exists() and wait_time < max_wait:
+                time.sleep(0.1)
+                wait_time += 0.1
+
+            if Path(metrics_file).exists():
+                with open(metrics_file, 'r') as f:
+                    metrics = json.load(f)
+            else:
+                print(f"Metrics file not found: {metrics_file}")
+                metrics = {}
 
             results.append({
                 "run_id": run_id,
@@ -184,20 +213,29 @@ def run_pysynthlab_experiments():
                 "file": str(sygus_file),
                 "return_code": retcode,
                 "time": elapsed_time,
+                "avg_cpu_percentage": avg_cpu_percentage, 
+                "max_memory_usage": max_memory,
+                "avg_all_cpu": json.dumps(avg_all_cpu),
                 "stdout": stdout.decode("utf-8"),
                 "stderr": stderr.decode("utf-8"),
                 **metrics
             })
 
             Path(temp_config_file).unlink()
-            Path(metrics_file).unlink()
+            if Path(metrics_file).exists():
+                Path(metrics_file).unlink()
 
-        write_csv_results(output_pysynthlab_csv, results, fieldnames)
+            write_csv_results(output_pysynthlab_csv, results, fieldnames)
         print(f"Finished writing results to {output_pysynthlab_csv}")
     print(f"All pysynthlab finished")
 
 
 if __name__ == "__main__":
-    for i in range(3):
-        #run_pysynthlab_experiments()
-        run_cvc5_experiments()
+    # for i in range(3):
+        # run_cvc5_experiments()
+    for i in range(1):
+        run_pysynthlab_experiments(30)
+    # for i in range(1):
+    #     run_pysynthlab_experiments(60)
+    # for i in range(1):
+    #     run_pysynthlab_experiments(120)
